Safari-609.1.11.1/LayoutTests/media/session/default-session.html

<!DOCTYPE html>
<html>
<head>
    <script src=../video-test.js></script>
    <script type="text/javascript">
        function doTest () 
        {
            var audioElement = document.createElement("audio");
            run("audioElement = document.createElement('audio')");
            testExpected("audioElement.session", null);

            endTest();
        }
    </script>
</head>
<body onload="doTest()">
    <p>New media elements are assigned to the Default media session, which is not exposed through the session attribute.</p>
</body>
</html>
